requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# UI
streamlit==1.28.0
plotly==5.18.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0

# Embeddings
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.2

# Vector Store
faiss-cpu==1.7.4

# Database (Optional - for metadata)
sqlalchemy==2.0.23
sqlite3  # Built-in

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1


